% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSRT.R
\name{compareSRT}
\alias{compareSRT}
\title{Summarize metrics for reference data and synthetic data}
\usage{
compareSRT(simsrt)
}
\arguments{
\item{simsrt}{A SRTsim object}
}
\value{
Returns an object with summarized metrics
}
\description{
Summarize metrics for reference data and synthetic data
}
