% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSRTLocs.R
\name{convert_grid}
\alias{convert_grid}
\title{Convert continuous coordinate into integer, essential for BayesSpace to determine the neighborhood info}
\usage{
convert_grid(x)
}
\arguments{
\item{x}{A numeric vector of continuous coordinate}
}
\value{
Returns a numeric vector oof integer coordinate
}
\description{
Convert continuous coordinate into integer, essential for BayesSpace to determine the neighborhood info
}
\examples{

## Create a simSRT object
toySRT  <- createSRT(count_in=toyData$toyCount,loc_in = toyData$toyInfo)
set.seed(1)

## Create New Locations within Profile
toySRT2  <- srtsim_newlocs(toySRT,new_loc_num=1000)

## Convert non-integer x-coordinates into an integer value
newGrid_x <- convert_grid(simcolData(toySRT2)$x)
}
