% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleLIBD}
\alias{exampleLIBD}
\title{Data used for creating vignettes}
\format{
A data list
\describe{
\item{example_count}{A sparse matrix with 80 rows and 3611 columns.}
\item{example_loc}{A data frame with 3611 rows and 6 columns.}
}
}
\source{
{created based on a SpatialLIBD data (SampleID: 151673) to serve as an example}
}
\usage{
exampleLIBD
}
\description{
A data list containing the a gene expression matrix and a location dataframe
}
\examples{
data(exampleLIBD)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
