% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTmodels.R
\name{fit_pos_optim}
\alias{fit_pos_optim}
\title{fitting data with poisson through optim function}
\usage{
fit_pos_optim(x, maxiter = 500)
}
\arguments{
\item{x}{A vector of count values to be fitted}

\item{maxiter}{number of iteration}
}
\value{
Returns a vector with mean paramter lambda, loglikelihood value llk, convergence
}
\description{
fitting data with poisson through optim function
}
