% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metrics.R
\name{get_metrics_pd}
\alias{get_metrics_pd}
\title{Extracted summarized metrics for reference data and synthetic data}
\usage{
get_metrics_pd(simsrt, metric = "GeneMean")
}
\arguments{
\item{simsrt}{A SRTsim object}

\item{metric}{Specification of metrics to be plotted.}
}
\value{
Returns a dataframe for ggplot
}
\description{
Extracted summarized metrics for reference data and synthetic data
}
