% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSRT.R
\name{get_stats_gene}
\alias{get_stats_gene}
\title{Summarize gene-wise summary metrics}
\usage{
get_stats_gene(mat, group, log_trans = TRUE)
}
\arguments{
\item{mat}{A count matrix}

\item{group}{A group label}

\item{log_trans}{A logical constant indicating whether to log transform the gene mean and variance}
}
\value{
Returns a n by 5 dataframe with location metrics
}
\description{
Summarize gene-wise summary metrics
}
