% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSRT.R
\name{get_stats_loc}
\alias{get_stats_loc}
\title{Summarize location-wise summary metrics}
\usage{
get_stats_loc(mat, group, log_trans = TRUE)
}
\arguments{
\item{mat}{A count matrix}

\item{group}{A group label}

\item{log_trans}{A logical constant indicating whether to log transform the libsize}
}
\value{
Returns a n by 3 dataframe with location metrics
}
\description{
Summarize location-wise summary metrics
}
