% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reGenCountshiny.R
\name{reGenCountshiny}
\alias{reGenCountshiny}
\title{ReSimulate Count Data with Parameters Specification from Shiny}
\usage{
reGenCountshiny(shinyOutput)
}
\arguments{
\item{shinyOutput}{A list of Shiny Output. Including a simCount, simInfo,simcountParam,simLocParam}
}
\value{
Returns a Count DataFrame
}
\description{
ReSimulate Count Data with Parameters Specification from Shiny
}
