% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTsim_class.R
\name{simcolData}
\alias{simcolData}
\title{Access synthetic colData}
\usage{
simcolData(x)
}
\arguments{
\item{x}{SRTsim object}
}
\value{
Returns the colData of synthetic data
}
\description{
Access synthetic colData
}
\examples{

## Create a simSRT object
toySRT  <- createSRT(count_in=toyData$toyCount,loc_in = toyData$toyInfo)
set.seed(1)
## Estimate model parameters for data generation
toySRT <- srtsim_fit(toySRT,sim_schem="tissue")
## Generate synthetic data with estimated parameters
toySRT <- srtsim_count(toySRT)
simcolData(toySRT)

}
