% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTsim_class.R
\name{simCounts}
\alias{simCounts}
\title{Access synthetic count matrix}
\usage{
simCounts(x)
}
\arguments{
\item{x}{SRTsim object}
}
\value{
Returns a synthetic count matrix
}
\description{
Access synthetic count matrix
}
\examples{

## Create a simSRT object
toySRT  <- createSRT(count_in=toyData$toyCount,loc_in = toyData$toyInfo)
set.seed(1)
## Estimate model parameters for data generation
toySRT <- srtsim_fit(toySRT,sim_schem="tissue")
## Generate synthetic data with estimated parameters
toySRT <- srtsim_count(toySRT)
simCounts(toySRT)[1:3,1:3]

}
