% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSRTLocs.R
\name{simNewLocs}
\alias{simNewLocs}
\title{Fit the marginal distributions for single gene}
\usage{
simNewLocs(newN, lay_out = c("grid", "random"), preLoc)
}
\arguments{
\item{newN}{A integer specifying the number of spatial locations in the synthetic data}

\item{lay_out}{A character string specifying arrangement of new generated spatial locations. Default is "grid"}

\item{preLoc}{A data frame of shape n by 3 that x, y coodinates and domain label}
}
\value{
Returns a n by 2 dataframe with newly generated spatial locations
}
\description{
Fit the marginal distributions for single gene
}
