% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTaffine.R
\name{srtsim_count_affine}
\alias{srtsim_count_affine}
\title{Generate Data with Estimated Parameters For A New Designed Pattern}
\usage{
srtsim_count_affine(
  simsrt,
  reflabel,
  targetlabel,
  breaktie = "random",
  nn_func = c("mean", "median", "ransam"),
  nn_num = 5,
  local_sid = NULL,
  numCores = 1
)
}
\arguments{
\item{simsrt}{A SRTsim object with estimated parameters from fitting step}

\item{reflabel}{A character vector specifying labels for reference regions}

\item{targetlabel}{A character vector specifying labels for target regions}

\item{breaktie}{A character string specifying how ties are treated.  Same as the "tie.method" in rank function}

\item{nn_func}{A character string specifying how the psedo-count to be generated. options include 'mean','median' and 'ransam'.}

\item{nn_num}{A integer of nearest neighbors, default is 5.}

\item{local_sid}{A numberic seed used locally for the affine transformation. Default is NULL.}

\item{numCores}{A number of cores to use}
}
\value{
Returns a SRTsim object with a newly generated count matrix
}
\description{
Generate Data with Estimated Parameters For A New Designed Pattern
}
