% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{SRTsim_shiny}
\alias{SRTsim_shiny}
\title{Run the SRTsim Shiny Application}
\usage{
SRTsim_shiny()
}
\value{
A list that contains a count matrix, a location dataframe, and all parameter specifications.
}
\description{
Run the SRTsim Shiny Application
}
