% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{A toyExample to showcase reference-based simulations}
\format{
A data list
\describe{
\item{toyCount}{A sparse matrix with 100 rows and 251 columns.}
\item{toyInfo}{A data frame with 251 rows and 3 columns.}
}
}
\source{
{created based on a ST Human Breast Cancer data to serve as an example}
}
\usage{
toyData
}
\description{
A data list containing the a gene expression matrix and a location dataframe
}
\examples{
data(toyData)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
