% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyShiny}
\alias{toyShiny}
\title{A toyExample to showcase reference-free simulations}
\format{
A list of shiny output
\describe{
\item{simCount}{A data frame with 150 rows and 980 columns.}
\item{simInfo}{A data frame with 980 rows and 4 columns: x, y, group, foldchange}
\item{simcountParam}{A list of user-specified parameters for count generation}
\item{simLocParam}{A list of user-specified parameters for pattern design}
}
}
\source{
{created based using the SRTsim_shiny()}
}
\usage{
toyShiny
}
\description{
A list of shiny output
}
\examples{
data(toyShiny)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
