% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSRTLocs.R
\name{convert_grid}
\alias{convert_grid}
\title{Convert continuous coordinate into integer, essential for BayesSpace to determine the neighborhood info}
\usage{
convert_grid(x)
}
\arguments{
\item{x}{A numeric vector of continuous coordinate}
}
\value{
Returns a numeric vector oof integer coordinate
}
\description{
Convert continuous coordinate into integer, essential for BayesSpace to determine the neighborhood info
}
