% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSRT.R
\name{createSRT}
\alias{createSRT}
\title{Create simSRT object}
\usage{
createSRT(count_in, loc_in, refID = "ref1")
}
\arguments{
\item{count_in}{A gene expression count \code{matrix}}

\item{loc_in}{A location \code{dataframe} with colnames x,y,label}

\item{refID}{A \code{character} reference sample identifier. Default = \code{ref1}.}
}
\value{
Returns a spatialExperiment-based object
}
\description{
Create simSRT object
}
\examples{

## Create a simSRT object
toySRT  <- createSRT(count_in=toyData$toyCount,loc_in = toyData$toyInfo)

## Explore the object
toySRT 
}
