% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleLIBD}
\alias{exampleLIBD}
\title{Data to showcase the simulation procedures}
\format{
A SRT data list
\describe{
\item{example_count}{A sparse matrix with 100 rows and 251 columns}
\item{example_loc}{A data frame with 251 rows and 3 columns:x,y,label}
}
}
\source{
{created based on a ST Human Breast Cancer data to serve as an example}
}
\usage{
exampleLIBD
}
\description{
A data list containing the a gene expression matrix and a
}
\examples{
data(exampleLIBD)       #Lazy loading. Data becomes visible as soon as called
}
\keyword{datasets}
