% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny2srt.R
\name{Shiny2SRT}
\alias{Shiny2SRT}
\title{Create a SRTsim object from reference-free shinyoutput}
\usage{
Shiny2SRT(shinyOutput)
}
\arguments{
\item{shinyOutput}{A list of Shiny Output. Including a simCount, simInfo,simcountParam,simLocParam}
}
\value{
Returns a SRTsim object with user-specified parameters stored in metaParam slot.
}
\description{
Create a SRTsim object from reference-free shinyoutput
}
\examples{

shinySRT <- Shiny2SRT(toyShiny)

## Explore the new SRT object
shinySRT@metaParam
shinySRT@simCounts[1:3,1:3]
shinySRT@simcolData
}
