% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSRT.R
\name{subsetSRT}
\alias{subsetSRT}
\title{Subset SRT object based on domain labels of interest}
\usage{
subsetSRT(simsrt, sel_label = NULL)
}
\arguments{
\item{simsrt}{A SRTsim object}

\item{sel_label}{A vector of selected domain labels used for the data generation}
}
\value{
Returns a spatialExperiment-based object
}
\description{
Subset SRT object based on domain labels of interest
}
\examples{

## Create a simSRT object
toySRT  <- createSRT(count_in=toyData$toyCount,loc_in = toyData$toyInfo)
set.seed(1)
## Only Keep the Spatial Locations labelled as "A" in the reference data
subtoySRT <- subsetSRT(toySRT,sel_label = "A")

}
